+++ server.log
@@ -1,0 +1,1 @@
+🚀 Server running at http://localhost:3000
